Started accounts authorization

+++ accounts.php
@@ -1,28 +1,43 @@
 <?php
 header("Access-Control-Allow-Origin: *");
 header("Access-Control-Allow-Headers: *");
+$allowedMethods= ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'];
 header("Access-Control-Allow-Methods: GET, POST, PUT, DELETE, OPTIONS");
 header("Content-Type: application/json; charset=UTF-8");
 header("Access-Control-Max-Age: 3600");
 
 require_once 'db_connection.php';
+require_once 'authorization.php';
 
-if ($_SERVER['REQUEST_METHOD'] === 'GET') {
-    handleGetRequest($conn);
-} elseif ($_SERVER['REQUEST_METHOD'] === 'POST') {
-    handlePostRequest($conn);
-} elseif ($_SERVER['REQUEST_METHOD'] === 'PUT') {
+$authResponse = authorizeUser();
+$auth = json_decode($authResponse, true);
+
+$user = null;
+$isAdmin = false;
+
+if (!isset($auth["message"])) {
+    $user = $auth['data'];
+    $isAdmin = $user['id'] == 1;
+}
+
+if ($_SERVER['REQUEST_METHOD'] === 'GET' && $user != null) {
+    handleGetRequest($conn, $user['id']);
+} elseif ($_SERVER['REQUEST_METHOD'] === 'POST' && $user != null) {
+    handlePostRequest($conn, $user['id']);
+} elseif ($_SERVER['REQUEST_METHOD'] === 'PUT' && $user != null) {
     handlePutRequest($conn);
-} elseif ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
+} elseif ($_SERVER['REQUEST_METHOD'] === 'DELETE' && $user != null) {
     handleDeleteRequest($conn);
 } elseif ($_SERVER['REQUEST_METHOD'] === 'OPTIONS') {
     handleOptionsRequest($conn);
-} else {
+} elseif (!in_array($_SERVER['REQUEST_METHOD'], $allowedMethods)) {
     sendJsonResponse(405, ["message" => "$_SERVER[REQUEST_METHOD] requests are not allowed"]);
+} else {
+    sendJsonResponse(403, ["message" => "You are not permitted to access this content!"]);
 }
 $conn->close();
 
-function handleGetRequest($conn) {
+function handleGetRequest($conn, $user_id) {
     global $accounts_table_name;
     if (isset($_GET['id'])) {
         $accountId = $_GET['id'];
@@ -33,12 +48,20 @@ function handleGetRequest($conn) {
 
         if ($result !== false && $result->num_rows > 0) {
             $account = $result->fetch_assoc();
+            if ($account['user_id'] != $user_id) {
+                sendJsonResponse(403, ['message'=> 'You are not permitted to access this account!']);
+                return;
+            }
             sendJsonResponse(200, $account);
         } else {
             sendJsonResponse(404, ["message" => 'Account not found']);
         }
         $stmt->close();
     } elseif (isset($_GET['user_id'])) {
+        if ($user_id != $_GET['user_id']) {
+            sendJsonResponse(403, ['message'=> 'You are not permitted to access this account!']);
+            return;
+        }
         $userId = $_GET['user_id'];
         $stmt = $conn->prepare("SELECT * FROM $accounts_table_name WHERE user_id = ?");
         $stmt->bind_param("i", $userId);
@@ -51,6 +74,9 @@ function handleGetRequest($conn) {
         }
         sendJsonResponse(200, $accounts);
     } else {
+        if ($user_id != 1) {
+            sendJsonResponse(403, ["message"=> "You are not allowed to get all accounts!"]);
+        }
         $result = $conn->query("SELECT * FROM $accounts_table_name");
         $accounts = [];
         while ($row = $result->fetch_assoc()) {
